+++ src/NeuralNet/CostFunctions/CostFunction.php
@@ -9,7 +9,7 @@ interface CostFunction
     /**
      * Return a tuple of the min and max output value for this function.
      *
-     * @return array
+     * @return float[]
      */
     public function range() : array;