+++ src/NeuralNet/CostFunctions/Exponential.php
@@ -44,7 +44,7 @@ class Exponential implements CostFunction
     /**
      * Return a tuple of the min and max output value for this function.
      *
-     * @return array
+     * @return float[]
      */
     public function range() : array
     {