+++ src/NeuralNet/CostFunctions/LeastSquares.php
@@ -19,7 +19,7 @@ class LeastSquares implements CostFunction
     /**
      * Return a tuple of the min and max output value for this function.
      *
-     * @return array
+     * @return float[]
      */
     public function range() : array
     {