+++ src/NeuralNet/CostFunctions/RelativeEntropy.php
@@ -43,7 +43,7 @@ class RelativeEntropy implements CostFunction
     /**
      * Return a tuple of the min and max output value for this function.
      *
-     * @return array
+     * @return float[]
      */
     public function range() : array
     {