+++ tests/Regressors/AdalineTest.php
@@ -9,7 +9,7 @@ use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Unlabeled;
 use Rubix\ML\Regressors\Adaline;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
-use Rubix\ML\NeuralNet\CostFunctions\LeastSquares;
+use Rubix\ML\NeuralNet\CostFunctions\HuberLoss;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
 use RuntimeException;
@@ -30,7 +30,7 @@ class AdalineTest extends TestCase
 
         $this->testing = $this->training->randomize()->head(3);
 
-        $this->estimator = new Adaline(100, 1, new Adam(0.01), 1e-4, new LeastSquares(), 1e-3);
+        $this->estimator = new Adaline(100, 1, new Adam(0.01), 1e-4, new HuberLoss(1.), 1e-3);
     }
 
     public function test_build_regressor()